Add: all movements logic and implementation

+++ account-service/src/main/java/projects/gsc/accountservice/dto/MovementCreateDto.java
@@ -12,5 +12,5 @@ public class MovementCreateDto {
     private double amount;
     private Long receiptAccNumber;
     private int ref;
-    private Long entity;
+    private int entity;
 }
